updated lexer
much more robust lexer, i defined a short dictionary to find keyword tokens and a short algorithm to define a string token. this is sort of  a cheat but this basically calls every word a variable until it gets sorted out later... but it works! I will also throw in a test doc i was using